+++ src/mokodolidymoindex.php
@@ -1,9 +1,9 @@
 <?php
 /* Copyright (C) 2025		Jonathan Miller				<jmiller@mokoconsulting.tech>
  *
- * This program is free software: you can redistribute it and/or modify
+ * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
+ * the Free Software Foundation; either version 3 of the License, or
  * (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
@@ -12,13 +12,13 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
 /**
- * \file    mokodolidymo/admin/about.php
- * \ingroup mokodolidymo
- * \brief   About page of module MokoDoliDymo.
+ *	\file       mokodolidymo/mokodolidymoindex.php
+ *	\ingroup    mokodolidymo
+ *	\brief      Home page of MokoDoliDymo module
  */
 
 // Load Dolibarr environment
@@ -40,6 +40,9 @@ if (!$res && $i > 0 && file_exists(substr($tmp, 0, ($i + 1))."/main.inc.php")) {
 if (!$res && $i > 0 && file_exists(dirname(substr($tmp, 0, ($i + 1)))."/main.inc.php")) {
 	$res = @include dirname(substr($tmp, 0, ($i + 1)))."/main.inc.php";
 }
+if (!$res && file_exists("../main.inc.php")) {
+	$res = @include "../main.inc.php";
+}
 if (!$res && file_exists("../../main.inc.php")) {
 	$res = @include "../../main.inc.php";
 }
@@ -50,10 +53,7 @@ if (!$res) {
 	die("Include of main fails");
 }
 
-// Libraries
-require_once DOL_DOCUMENT_ROOT.'/core/lib/admin.lib.php';
-require_once DOL_DOCUMENT_ROOT.'/core/lib/functions2.lib.php';
-require_once '../lib/mokodolidymo.lib.php';
+require_once DOL_DOCUMENT_ROOT.'/core/class/html.formfile.class.php';
 
 /**
  * @var Conf $conf
@@ -63,17 +63,18 @@ require_once '../lib/mokodolidymo.lib.php';
  * @var User $user
  */
 
-// Translations
-$langs->loadLangs(array("errors", "admin", "mokodolidymo@mokodolidymo"));
+// Load translation files required by the page
+$langs->loadLangs(array("mokodolidymo@mokodolidymo"));
 
-// Access control
-if (!$user->admin) {
+$action = GETPOST('action', 'aZ09');
+
+// Security check
+if (!isModEnabled('mokodolidymo')) {
+	accessforbidden('Module not enabled');
+}
+if (!$user->hasRight('mokodolidymo', 'label', 'read')) {
 	accessforbidden();
 }
-
-// Parameters
-$action = GETPOST('action', 'aZ09');
-$backtopage = GETPOST('backtopage', 'alpha');
 
 
 /*
@@ -82,25 +83,16 @@ $backtopage = GETPOST('backtopage', 'alpha');
 
 $form = new Form($db);
 
-$help_url = '';
-$title = "MokoDoliDymoSetup";
+llxHeader("", $langs->trans("MokoDoliDymoArea"), '', '', 0, 0, '', '', '', 'mod-mokodolidymo page-index');
 
-llxHeader('', $langs->trans($title), $help_url, '', 0, 0, '', '', '', 'mod-mokodolidymo page-admin_about');
+print load_fiche_titre($langs->trans("MokoDoliDymoArea"), '', 'fa-print');
 
-// Subheader
-$linkback = '<a href="'.($backtopage ? $backtopage : DOL_URL_ROOT.'/admin/modules.php?restore_lastsearch_values=1').'">'.$langs->trans("BackToModuleList").'</a>';
+print '<div class="fichecenter">';
+print '<div class="fichethirdleft">';
+print '</div>';
+print '<div class="fichetwothirdright">';
+print '</div>';
+print '</div>';
 
-print load_fiche_titre($langs->trans($title), $linkback, 'title_setup');
-
-// Configuration header
-$head = mokodolidymoAdminPrepareHead();
-print dol_get_fiche_head($head, 'about', $langs->trans($title), 0, 'fa-print');
-
-dol_include_once('/mokodolidymo/core/modules/modMokoDoliDymo.class.php');
-$tmpmodule = new modMokoDoliDymo($db);
-print $tmpmodule->getDescLong();
-
-// Page end
-print dol_get_fiche_end();
 llxFooter();
 $db->close();